requirements.txt
boto3
awscli
numpy
pyyaml==5.1